feat: provera uloge korisnika

+++ src/Explorer.API/Controllers/QuizController.cs
@@ -3,27 +3,24 @@ using AutoMapper;
 using Explorer.Tours.API.Dtos;
 using Explorer.Tours.API.Public;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+
 
 namespace Explorer.API.Controllers
 {
+    [Authorize(Roles = "Author")]
     [Route("api/quizzes")]
     [ApiController]
     public class QuizController : ControllerBase
     {
         private readonly IQuizService _quizService;
         private readonly IMapper _mapper;
-        private IQuizService quizService;
 
         public QuizController(IQuizService quizService, IMapper mapper)
         {
             _quizService = quizService;
             _mapper = mapper;
 
-        }
-
-        public QuizController(IQuizService quizService)
-        {
-            this.quizService = quizService;
         }
 
         [HttpGet("{id}")]
@@ -40,18 +37,34 @@ namespace Explorer.API.Controllers
         [HttpPost]
         public ActionResult<QuizDto> Create([FromBody] CreateQuizDto quiz)
         {
+            quiz.AuthorId = GetUserIdFromToken();
             return Ok(_quizService.Create(quiz));
+        }
+
+        private long GetUserIdFromToken()
+        {
+            return long.Parse(User.FindFirst("id").Value);
         }
 
         [HttpPut("{id}")]
         public ActionResult<QuizDto> Update(long id, [FromBody] QuizDto quiz)
         {
+            var existing = _quizService.Get(id);
+
+            if (existing.AuthorId != GetUserIdFromToken())
+                return Forbid(); // 403
+
             return Ok(_quizService.Update(quiz, id));
         }
 
         [HttpDelete("{id}")]
         public IActionResult Delete(long id)
         {
+            var existing = _quizService.Get(id);
+
+            if (existing.AuthorId != GetUserIdFromToken())
+                return Forbid();
+
             _quizService.Delete(id);
             return NoContent();
         }
@@ -60,20 +73,25 @@ namespace Explorer.API.Controllers
         [HttpPost("{quizId}/questions")]
         public ActionResult<QuizDto> AddQuestion(long quizId, [FromBody] CreateQuestionDto question)
         {
-            if (question == null)
-                return BadRequest("Question cannot be null.");
+            var quiz = _quizService.Get(quizId);
 
-           // var questionDto = _mapper.Map<QuestionDto>(question);
-            var updatedQuiz = _quizService.AddQuestion(quizId, question);
-            return Ok(updatedQuiz);
+            if (quiz.AuthorId != GetUserIdFromToken())
+                return Forbid();
+
+            return Ok(_quizService.AddQuestion(quizId, question));
         }
 
         [HttpDelete("{quizId}/questions/{questionId}")]
         public ActionResult<QuizDto> RemoveQuestion(long quizId, long questionId)
         {
-            var updatedQuiz = _quizService.RemoveQuestion(quizId, questionId);
-            return Ok(updatedQuiz);
+            var quiz = _quizService.Get(quizId);
+
+            if (quiz.AuthorId != GetUserIdFromToken())
+                return Forbid();
+
+            return Ok(_quizService.RemoveQuestion(quizId, questionId));
         }
+
 
     }
 }
